Fix a major issue with reliable message acking
Nothing was advancing the "most recently acked message" which meant that
connections would eventually die due to trying to replay every message
to the client.

Things seem to be mostly working now, only issue I can see is that
reconnecting doesn't seem to work without rebooting the client.

+++ common/d_client.h
@@ -76,7 +76,7 @@ class SVCMessages
 
 	uint32_t m_nextPacketID;
 	uint16_t m_nextReliableID;
-	uint32_t m_reliableNoAck;
+	uint16_t m_reliableNoAck;
 
 	sentPacket_s& sentPacket(const uint32_t id);
 	sentPacket_s* validSentPacket(const uint32_t id);
